+++ src/app/components/create-qrcode/components/settingUser/create-qrcode.component.scss
@@ -99,7 +99,7 @@
     box-shadow: 0px 5px 20px rgba(49, 52, 61, 0.05);
     border-radius: 10px;
 
-    padding: 20px 30px;
+    // padding: 20px 30px;
 
     & .sectionBorder {
       background: #e1e3e1;
