+++ src/app/components/input-user-tips/input-user-tips.component.scss
@@ -3,7 +3,6 @@
   --content: none
   display: block;
   width: 100%;
-  --borderColor:#c8c9cf
 
 }
 
@@ -36,13 +35,16 @@ input[type="number"] {
 
     background: #ffffff;
     border: 1px solid #c8c9cf;
-    border-color: var(--borderColor);
+    border-color: #c8c9cf;
     border-radius: 6px;
     flex-grow: 1;
     flex-shrink: 0;
 
     padding: 25px;
     margin-bottom: 20px;
+  }
+  & .inValid {
+    border-color: #f13608;
   }
 }
