+++ src/app/components/table/table.component.ts
@@ -3,7 +3,6 @@ import {
   ChangeDetectionStrategy,
   ChangeDetectorRef,
   Component,
-  DestroyRef,
   inject,
 } from '@angular/core';
 import { Observable } from 'rxjs';
@@ -29,7 +28,7 @@ export class TableComponent {
   readonly #inputService = inject(switchOnService);
   readonly #filterService = inject(SortDataService);
   readonly #cdr = inject(ChangeDetectorRef);
-  readonly #destroyRef = inject(DestroyRef);
+  // readonly #destroyRef = inject(DestroyRef);
 
   public btnText: ButtonData = {
     text: 'Скачать в Exel',
