+++ src/app/components/upload-logo/upload-logo.component.ts
@@ -1,18 +1,16 @@
 import {
-  AfterViewInit,
   ChangeDetectionStrategy,
   Component,
   DestroyRef,
   ElementRef,
-  EventEmitter,
   inject,
   OnInit,
-  Output,
   ViewChild,
 } from '@angular/core';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 import { SharedModule } from '../../shared.module';
 import { ButtonService } from '../buttons/service/buttons.component.service';
+import { UploadTransmitPhotoService } from '../create-qrcode/components/services/uploadTransmitPhoto.service';
 import { ButtonData } from './../../types/sectionItem';
 
 @Component({
@@ -22,10 +20,10 @@ import { ButtonData } from './../../types/sectionItem';
   styleUrl: './upload-logo.component.scss',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class UploadLogoComponent implements AfterViewInit, OnInit {
+export class UploadLogoComponent implements OnInit {
   @ViewChild('fileInput') input?: ElementRef;
-  @ViewChild('customInput') customInput?: ElementRef;
-  @Output() uploadedFile = new EventEmitter<File>();
+  // @ViewChild('customInput') customInput?: ElementRef;
+  // @Output() uploadedFile = new EventEmitter<File>();
 
   btnText: ButtonData = {
     iconClass: 'icon-icon-upload',
@@ -39,30 +37,22 @@ export class UploadLogoComponent implements AfterViewInit, OnInit {
 
   readonly #clickOnBTN = inject(ButtonService);
   readonly #destroyRef = inject(DestroyRef);
+  readonly #transmitLogoService = inject(UploadTransmitPhotoService);
 
   ngOnInit(): void {
     this.#clickOnBTN.eventClick$
       .pipe(takeUntilDestroyed(this.#destroyRef))
       .subscribe((data) => {
         if (data.id === 6) {
-          console.log('Click on Input');
-
           this.input?.nativeElement.click();
         }
       });
   }
-
-  ngAfterViewInit(): void {
-    // console.log(this.input);
-  }
-
   openFile() {
     const file = this.input?.nativeElement.files[0];
-    // console.log(typeof file);
 
-    console.log(file);
     if (file) {
-      this.uploadedFile.emit(file);
+      this.#transmitLogoService.getPhotoFromComponent(file);
     }
   }
 }
